answer script check

+++ database/migrations/2021_02_04_155803_create_score_to_answers.php
@@ -14,7 +14,7 @@ class CreateScoreToAnswers extends Migration
     public function up()
     {
         Schema::table('answers', function (Blueprint $table) {
-            $table->integer('score')->default(0);
+            $table->float('score',8,2)->default(0);
         });
     }
